refactor(ui): move semver utils
