+++ app/react/portainer/environments/update-schedules/common/utils.ts
@@ -1,26 +1,4 @@
-/**
- * Compares two semver strings.
- *
- * returns:
- * - `-1` if `a < b`
- * - `0` if `a == b`
- * - `1` if `a > b`
- */
-export function semverCompare(a: string, b: string) {
-  if (a.startsWith(`${b}-`)) {
-    return -1;
-  }
-
-  if (b.startsWith(`${a}-`)) {
-    return 1;
-  }
-
-  return a.localeCompare(b, undefined, {
-    numeric: true,
-    sensitivity: 'case',
-    caseFirst: 'upper',
-  });
-}
+import { semverCompare } from '@/react/common/semver-utils';
 
 export function compareVersion(
   currentVersion: string,
