+++ app/react/portainer/environments/update-schedules/queries/useSupportedAgentVersions.ts
@@ -2,8 +2,7 @@ import { useQuery } from 'react-query';
 
 import axios, { parseAxiosError } from '@/portainer/services/axios';
 import { withError } from '@/react-tools/react-query';
-
-import { semverCompare } from '../common/utils';
+import { semverCompare } from '@/react/common/semver-utils';
 
 import { queryKeys } from './query-keys';
 import { buildUrl } from './urls';
